# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rawmart Blog API",
        "description": "API documentation for the RAWMART BLOG ASSESSMENT API",
        "contact": {
            "email": "ibrahimsharafadeen95@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/posts/{post}/comments": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Add comment to a post",
                "operationId": "f2a2f91d5891b6a3755e1350158646c8",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "Nice post!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/{post}/comments/{commentId}": {
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Update a comment",
                "operationId": "e1462355c23dffd0b96f5ee3eb63bb61",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "Nice post!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comment updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comment not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Delete a comment",
                "operationId": "7ef646ddd850623630107dd4260fae6d",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Comment deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserSummary": {
                "description": "Minimal user representation",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "new-user"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "image": {
                        "type": "string",
                        "example": "https://images.unsplash.com/photo-1535713875002-d1d0cf377fde"
                    }
                },
                "type": "object"
            },
            "Comment": {
                "description": "Comment resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 101
                    },
                    "time": {
                        "type": "string",
                        "example": "3 minutes ago"
                    },
                    "user_id": {
                        "type": "string",
                        "example": "new-user"
                    },
                    "text": {
                        "type": "string",
                        "example": "Great insights!"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserSummary"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}